requirements.txt
pandas
plotly
pygit
pathlib
requests
ccscanner==0.1.10
